Bug Fix: The list of Project Users was not being used at all.  If a project lists users than only those users will see the new color display.
Updates: Minor code refactoring.

New: Another color combo defined in the README.md

Minor change to debugging: Previously if users were listed and the user was not a specified user than no debug info would be outputed.  Now even if it is not a specified user the debug info will still show up in the console.